01_ai_agents_first/02_openrouter/OpenRouter_Ai_agent_sdk/pyproject.toml
[project]
name = "openrouter-ai-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
]
